requirements.txt
Django==5.2.6
django-configurations==2.5.1
pillow==11.3.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-cors-headers==4.9.0
drf-standardized-errors==0.15.0
python-decouple==3.8
arrow==1.3.0